Fixed the code so it work with C++11

+++ TrikytaEngine/src/core/Common/EngineInstance.h
@@ -42,8 +42,8 @@ public:
 
 	inline void setEngineState(bool p_State) { m_EngineState = p_State; }
 
-	const unsigned int GetScreenWidth() const { return m_EngineConfig.WINDOW_WIDTH; }
-	const unsigned int GetScreenHeight() const { return m_EngineConfig.WINDOW_HEIGHT; }
+	const int GetScreenWidth() const { return (int)m_EngineConfig.WINDOW_WIDTH; }
+	const int GetScreenHeight() const { return (int)m_EngineConfig.WINDOW_HEIGHT; }
 
 private:
 	struct ENGINE_CONFIG& m_EngineConfig;
